Refactor email schedule form state management
