+++ packages/ui/src/components/cms/marketing/emailScheduling/index.ts
@@ -1,4 +1,7 @@
 export * from "./EmailScheduleForm";
 export * from "./EmailSchedulePreviewPanel";
 export * from "./EmailScheduleSummaryCard";
+export * from "./FollowUpControls";
+export * from "./ScheduleTimingFields";
+export * from "./hooks";
 export * from "./types";
